requirements.txt
google-genai==1.66.0
python-dotenv>=1.2.2
protobuf==7.34.0
cffi==2.0.0
requests==2.32.5
cryptography==46.0.5
google-auth==2.49.0